refactor: Collectors are now in their own folder
