+++ esmExport.js
@@ -60,8 +60,8 @@ export { default as Resolver } from './src/Libraries/definitions/Resolver';
 // Discord
 export { default as Embed } from './src/Utility/Discord/Embed';
 export { default as Prompt } from './src/Utility/Discord/Prompt';
-export { default as MessageCollector } from './src/Utility/Discord/MessageCollector';
-export { default as ReactionCollector } from './src/Utility/Discord/ReactionCollector';
+export { default as MessageCollector } from './src/Utility/Discord/Collectors/MessageCollector';
+export { default as ReactionCollector } from './src/Utility/Discord/Collectors/ReactionCollector';
 // External
 export { default as Stack } from './src/Utility/External/Stack';
 export { default as Queue } from './src/Utility/External/Queue';
